Merge rebrand into main

+++ Shared/CoreDataStack.swift
@@ -10,49 +10,70 @@ import CoreData
 
 class CoreDataStack {
     static let shared = CoreDataStack()
+    private let storageType: StorageType
 
-    private init() {}
+    init(_ storageType: StorageType = .persistent) {
+        self.storageType = storageType
 
-    // Modified from https://stackoverflow.com/a/57020353
+        if case .inmemory(let sampleData) = storageType {
+            self.initializeWithSampleData(with: sampleData)
+        }
+    }
+
     lazy var persistentContainer: NSPersistentContainer = {
-        let container = NSPersistentContainer(name: "Stockpile")
-        let storeURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.io.github.taylor-young0.Stockpile")!.appendingPathComponent("Stockpile.sqlite")
+        if case .inmemory(_) = storageType {
+            let container = NSPersistentContainer(name: "Stockpile")
+            let description = NSPersistentStoreDescription(url: URL(fileURLWithPath: "/dev/null"))
+            container.persistentStoreDescriptions = [description]
 
-        var defaultURL: URL?
-        if let storeDescription = container.persistentStoreDescriptions.first, let url = storeDescription.url {
-            defaultURL = FileManager.default.fileExists(atPath: url.path) ? url : nil
-        }
-
-        if defaultURL == nil {
-            container.persistentStoreDescriptions = [NSPersistentStoreDescription(url: storeURL)]
-        }
-        container.loadPersistentStores(completionHandler: { [unowned container] (storeDescription, error) in
-            if let error = error as NSError? {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
+            container.loadPersistentStores { _, error in
+                if let error = error as NSError? {
+                    fatalError("Unresolved error \(error), \(error.userInfo)")
+                }
             }
 
-            if let url = defaultURL, url.absoluteString != storeURL.absoluteString {
-                let coordinator = container.persistentStoreCoordinator
-                if let oldStore = coordinator.persistentStore(for: url) {
-                    do {
-                        try coordinator.migratePersistentStore(oldStore, to: storeURL, options: nil, withType: NSSQLiteStoreType)
-                    } catch {
-                        print(error.localizedDescription)
-                    }
+            return container
+        } else {
+            // Modified from https://stackoverflow.com/a/57020353
+            let container = NSPersistentContainer(name: "Stockpile")
+            let storeURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.io.github.taylor-young0.Stockpile")!.appendingPathComponent("Stockpile.sqlite")
 
-                    // delete old store
-                    let fileCoordinator = NSFileCoordinator(filePresenter: nil)
-                    fileCoordinator.coordinate(writingItemAt: url, options: .forDeleting, error: nil, byAccessor: { url in
+            var defaultURL: URL?
+            if let storeDescription = container.persistentStoreDescriptions.first, let url = storeDescription.url {
+                defaultURL = FileManager.default.fileExists(atPath: url.path) ? url : nil
+            }
+
+            if defaultURL == nil {
+                container.persistentStoreDescriptions = [NSPersistentStoreDescription(url: storeURL)]
+            }
+            container.loadPersistentStores(completionHandler: { [unowned container] (storeDescription, error) in
+                if let error = error as NSError? {
+                    fatalError("Unresolved error \(error), \(error.userInfo)")
+                }
+
+                if let url = defaultURL, url.absoluteString != storeURL.absoluteString {
+                    let coordinator = container.persistentStoreCoordinator
+                    if let oldStore = coordinator.persistentStore(for: url) {
                         do {
-                            try FileManager.default.removeItem(at: url)
+                            try coordinator.migratePersistentStore(oldStore, to: storeURL, options: nil, withType: NSSQLiteStoreType)
                         } catch {
                             print(error.localizedDescription)
                         }
-                    })
+
+                        // delete old store
+                        let fileCoordinator = NSFileCoordinator(filePresenter: nil)
+                        fileCoordinator.coordinate(writingItemAt: url, options: .forDeleting, error: nil, byAccessor: { url in
+                            do {
+                                try FileManager.default.removeItem(at: url)
+                            } catch {
+                                print(error.localizedDescription)
+                            }
+                        })
+                    }
                 }
-            }
-        })
-        return container
+            })
+            return container
+        }
     }()
 }
